requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
python-dotenv==0.16.0
psycopg2==2.8.6